The draft's autosave and flush move behind useComposerDraft
Title, body, the 2s debounce, and the cancel-then-write flush are one concern:
what has and has not reached the server. The last-persisted marks now live
where the autosave effect lives and are never read from outside — callers ask
isDirty() / needsTitle() instead, so no other file can drift on what unsaved
means. flushEdits/hasUnsavedEdits/draftNeedsTitle are re-exported for the
direct unit tests that prove the ordering.
